+++ CHANGELOG.md
@@ -172,6 +172,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       when passed to the command ([issue #1837](https://github.com/dsccommunity/SqlServerDsc/issues/1837)).
     - Now returns a more clear error message when the status of a database
       instance is not `Online`.
+- SqlTraceFlag
+  - The examples was updated to show that values should be passed as an array,
+    even when there is only one value.
+  - `Get-TargetResource` was updated to always return an array for parameter
+    `TraceFlags`.
 
 ### Fixed